Convert cached status codes into char lists
Currently, for the ibrowse adapter, status codes are converted back into
char lists when loaded from a json file, but not when they're pulled
back out of the cache. If you cache a request and then immediately try
to use it (without writing it and reading it from a file) the status
code we get back from the ibrowse call will be an integer. This fixes
that and transforms all status codes back into char lists when pulled
directly from the cache.