Fix another issue with hackney and `:with_body`.
`ExVCR.Adapter.Hackney.Converter.response_to_String` assumes
that the 4th tuple element is a client ref, and inspects it
as a form of getting a string body. However, when using the
`:with_body`, this caused the body to be quoted during playback
(e.g. `"foo"` instead of `foo`) since that's what `inspect` does
when given a string.

The solution is to see what the 4th element is and treat it appropriately.