Fix ignore localhost when using headers
Fixes the ignore_localhost option when recording requests that send
headers. This was raising an ArgumentError because every element of the
request was being enumerated and checked as a String (including
non-string-like maps, which is how headers are represented).

This changes the checking of the url to get the url out of the request
via the corresponding adapter.