Add a test case demonstrating an error with the ignore_localhost option when headers are set
The ignore_localhost check tries to convert each request property to a string to check for localhost,
which causes an error if there is a header keyword list specified.

Note this error is not reproducible for localhost URLs as the URL is always the first field,
thus passing the Enum.any? check before trying to test the header list.