Make global mock experimental feature and disable it by default
Even though all tests passed and the global mock was tested with an external application,
it was pretty optimistic to enabled that by default.
This commit changes default behavior that mocks every cassette and adds an ability to enable
the experimental feature. To do so, extend your `test.exs`:

```elixir
config :exvcr, [
  global_mock: true
]
```

In Travis CI we run the same test for both default behaviour and global
mock.