fix: link rendering and code highlights
The readme has some links that are not rendering well on
https://hexdocs.pm/exvcr, in addition some of the code examples use
`Elixir` as language name, so the highlighting doesn't work.

I've changed the links so that they render well on hexdocs and rexpect
the existing `target="_blank"`.

I've also changed the language in the codeblocks so that it is correctly
highlighted on both hexdocs and GitHub.